pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]

[project]
name = "ohgw-experiments"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "tqdm",
    "pyyaml",
    "pandas",
    "matplotlib",
    "seaborn",
    "scikit-learn",
]
